+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "cmake.sourceDirectory": "E:/CODE/CloudSeal/gRPC-Client"
+}
