+++ gRPC-Client/resources/templates/_common_behavior_properties.inja
@@ -1,0 +1,4 @@
+enabled: {{ enabled | default(true) }}
+focusPolicy: {{ focusPolicy | default("Qt.StrongFocus") }}
+hoverEnabled: {{ hoverEnabled | default(true) }}
+tooltip: {{ tooltip | default("""") }}
