+++ gRPC-Client/src/main.qml
@@ -1,0 +1,17 @@
+import QtQuick 2.9
+import QtQuick.Window 2.2
+
+Window {
+    visible: true
+    width: 640
+    height: 480
+    title: "CloudSealQt"
+    Text {
+        anchors.fill: parent
+        horizontalAlignment: Text.AlignHCenter
+        verticalAlignment: Text.AlignVCenter
+        font.bold: true
+        font.pointSize: 42
+        text: "Hello World!"
+    }
+}
